More work on date and settings
